requirements.txt
click==8.0.4
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
PyMySQL==1.0.2
PyYAML==6.0
SQLAlchemy==1.4.32
Werkzeug==2.0.3
